+++ config.yaml
@@ -1,16 +1,12 @@
 
 paths:
   raw_dir: "data/raw/*.json"
-  raw_abuseipdb: "data/raw/abuseipdb.json"
   raw_urlhaus: "data/raw/urlhaus.json"
   raw_malwarebazaar: "data/raw/malwarebazaar.json"
+  raw_feodo: "data/raw/feodotracker.json"
   normalized: "data/normalized/indicators.json"
   docs_coverage: "docs/attack-coverage.json"
   sigma_output: "detections/sigma/all_rules.yml"
   suricata_output: "detections/suricata/all_rules.rules"
   kql_output: "detections/kql/all_rules.kql"
   mappings: "mappings.json"
-
-api_keys:
-  abuseipdb: "ABUSEIPDB_API_KEY"
-  otx: "OTX_API_KEY"
