+++ data/raw/feodotracker.json
@@ -1,0 +1,56 @@
+[
+  {
+    "type": "ip",
+    "value": "162.243.103.246",
+    "source": "FeodoTracker",
+    "tags": [
+      "Emotet"
+    ],
+    "first_seen": "2026-01-28"
+  },
+  {
+    "type": "ip",
+    "value": "167.86.75.145",
+    "source": "FeodoTracker",
+    "tags": [
+      "Emotet"
+    ],
+    "first_seen": "2026-01-28"
+  },
+  {
+    "type": "ip",
+    "value": "137.184.9.29",
+    "source": "FeodoTracker",
+    "tags": [
+      "QakBot"
+    ],
+    "first_seen": "2026-01-28"
+  },
+  {
+    "type": "ip",
+    "value": "50.16.16.211",
+    "source": "FeodoTracker",
+    "tags": [
+      "QakBot"
+    ],
+    "first_seen": "2026-01-28"
+  },
+  {
+    "type": "ip",
+    "value": "34.204.119.63",
+    "source": "FeodoTracker",
+    "tags": [
+      "QakBot"
+    ],
+    "first_seen": "2026-01-28"
+  },
+  {
+    "type": "ip",
+    "value": "172.232.59.14",
+    "source": "FeodoTracker",
+    "tags": [
+      "QakBot"
+    ],
+    "first_seen": "2026-01-28"
+  }
+]
